test(ui): colocate washText coverage with styles.go
Move TestWashTextFitsContent next to the helper it exercises.
